interactive: moves the rewrite to pass conversion from app.py into get_all_possibl… (#2187)
moves the rewrite to pass conversion from app.py into
get_all_possible_rewrites
